Add cov2D model and list comparing the two current models
One annoying thing that took me a long time was in order to get the cov2d model working, I needed to change

Cov_model.compile(loss='sparse_categorical_crossentropy',
                      optimizer="adam",
                      metrics=['accuracy'])

To :

Cov_model.compile(loss='categorical_crossentropy',
                      optimizer="adam",
                      metrics=['accuracy'])

Which was annoying because python was throwing errors about expecting something about an output? of 1 and it was receiving an output? of 10.
Because of this I thought the model was having issues with either the test and training
data, or the inputs/outputs of the model. Very Frustrating, as it took me all last evening and this morning to noon to figure out.

And now with this commit I find I have to run each model independtly. Or else the conv2D model will fail after its first epoch.